pat parsing
