+++ common/src/mpegts/payload.rs
@@ -1,5 +1,6 @@
 use serde::{Deserialize, Serialize};
-use crate::{pes::PesPacketHeader, psi::PsiTypes};
+use crate::mpegts::pes::PesPacketHeader;
+use crate::mpegts::psi::PsiTypes;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub enum PayloadType {
